# HuyShopProductService/src/main/resources/openapi.yml
openapi: 3.0.1
info:
  title: HuyShop API
  version: 1.0.0

tags:
  - name: HuyShop
    description: HUY SHOP API SERVER

paths:
  /category/all:
    get:
      tags:
        - HuyShop
      summary: ALL CATEGORY
      operationId: getAllCategory
      responses:
        200:
          description: Ok
          content:
            application/json:
              schema:
                items:
                  type: array
                  $ref: '#/components/schemas/Category'
        400:
          description: Invalid ID supplied
          content: {}
        404:
          description: Category not found
          content: {}
        405:
          description: Validation exception
          content: {}

  /category:
    post:
      tags:
        - HuyShop
      summary: CREATE NEW CATEGORY
      operationId: createCategory
      requestBody:
        description: create new category
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Category'
          application/xml:
            schema:
              $ref: '#/components/schemas/Category'
        required: true
      responses:
        200:
          description: Ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
        400:
          description: Invalid ID supplied
          content: {}
        404:
          description: Category not found
          content: {}
        405:
          description: Validation exception
          content: {}
  /category/{id}:
    put:
      tags:
        - HuyShop
      summary: UPDATE CATEGORY
      operationId: updateCategory
      parameters:
        - name: id
          in: path
          description: ID of category to return
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: update category
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Category'
        required: true
      responses:
        200:
          description: Ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
        400:
          description: Invalid ID supplied
          content: {}
        404:
          description: Category not found
          content: {}
        405:
          description: Validation exception
          content: {}
    get:
      tags:
        - HuyShop
      summary: FIND CATEGORY BY ID
      operationId: findById
      parameters:
        - name: id
          in: path
          description: ID of category to return
          required: true
          schema:
            type: integer
            format: int64
      responses:
        200:
          description: Ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
        400:
          description: Invalid ID supplied
          content: {}
        404:
          description: Category not found
          content: {}
        405:
          description: Validation exception
          content: {}


    delete:
      tags:
        - HuyShop
      summary: DELETE CATEGORY EXITS
      operationId: deleteCategory
      parameters:
        - name: id
          in: path
          description: ID of category to return
          required: true
          schema:
            type: integer
            format: int64
      responses:
        200:
          description: successful operation
        400:
          description: Invalid ID supplied
          content: {}
        404:
          description: Category not found
          content: {}

  /product/all:
    get:
      tags:
        - HuyShop
      summary: ALL PRODUCT
      operationId: getAllProduct
      responses:
        200:
          description: Ok
          content:
            application/json:
              schema:
                items:
                  type: array
                  $ref: '#/components/schemas/Product'
        400:
          description: Invalid ID supplied
          content: {}
        404:
          description: Product not found
          content: {}
        405:
          description: Validation exception
          content: {}

  /product:
    post:
      tags:
        - HuyShop
      summary: CREATE NEW PRODUCT
      operationId: createProduct
      requestBody:
        description: create new product
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Product'
        required: true
      responses:
        200:
          description: Ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        400:
          description: Invalid ID supplied
          content: {}
        404:
          description: Category not found
          content: {}
        405:
          description: Validation exception
          content: {}

  /product/{id}:
    put:
      tags:
        - HuyShop
      summary: UPDATE CATEGORY
      operationId: updateProduct
      parameters:
        - name: id
          in: path
          description: ID of category to return
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: update product
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Product'
        required: true
      responses:
        200:
          description: Ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        400:
          description: Invalid ID supplied
          content: {}
        404:
          description: Category not found
          content: {}
        405:
          description: Validation exception
          content: {}
    get:
      tags:
        - HuyShop
      summary: FIND PRODUCT BY ID
      operationId: findByProductId
      parameters:
        - name: id
          in: path
          description: ID of product to return
          required: true
          schema:
            type: integer
            format: int64
      responses:
        200:
          description: Ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        400:
          description: Invalid ID supplied
          content: {}
        404:
          description: Category not found
          content: {}
        405:
          description: Validation exception
          content: {}

    delete:
      tags:
        - HuyShop
      summary: DELETE PRODUCT EXITS
      operationId: deleteProduct
      parameters:
        - name: id
          in: path
          description: ID of product to return
          required: true
          schema:
            type: integer
            format: int64
      responses:
        200:
          description: successful operation
        400:
          description: Invalid ID supplied
          content: {}
        404:
          description: Product not found
          content: {}

components:

  schemas:
    Category:
      description: This is CategoryEntity Model
      type: object
      required:
        - id
      properties:
        id:
          type: integer
          format: int64
        name:
          description: The name of CategoryEntity
          type: string
        status:
          type: boolean
    Product:
      description: This is CategoryEntity Model
      type: object
      required:
        - id
      properties:
        id:
          type: integer
          format: int64
        name:
          description: The name of CategoryEntity
          type: string
        status:
          type: boolean
        quantity:
          type: integer
          format: int64
        price:
          type: integer
          format: int64
        categoryId:
          type: Category